+++ CHANGELOG.md
@@ -5,7 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [ v0.1.0 ]
+## [ v0.1.1 ] - 2025-08-27
+
+### Changed
+- Logger to a singleton, and added respective methods
+
+### Added
+- Option to save logs to a report file
+- Method / logic to determine path to the file calling the method
+
+### Removed
+- Method determining the initializer file
+
+## [ v0.1.0 ] - 2025-08-22
 ### Changed
 - Cleaning up python syntax, adding annotations, and updating comments.
 
@@ -17,6 +29,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 - Deprecating methods for name changes and for refactoring
 
-## [ v0.0.1 ]
+## [ v0.0.1 ] - 2024-03-15
 ### Added
 - Changelog to mark the start of versioning.